requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
telethon==1.32.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0 